fix(core-database): decouple database package so that it has using it introduces no extra dependencies
