--- packages/core-database/src/actions/index.ts
@@ -1,1 +1,0 @@
-export { GetActiveDelegatesAction } from "./get-active-delegates";
