+++ packages/core-kernel/src/contracts/p2p/server.ts
@@ -19,6 +19,7 @@ export interface CurrentRound {
     nextForger: DelegateWallet;
     lastBlock: Interfaces.IBlockData;
     canForge: boolean;
+    timeLeftInSlot: number;
 }
 
 export interface ForgingTransactions {
