+++ packages/core-state/src/actions/index.ts
@@ -1,1 +1,2 @@
 export { BuildDelegateRankingAction } from "./build-delegate-ranking";
+export { GetActiveDelegatesAction } from "./get-active-delegates";
