+++ packages/core-state/src/index.ts
@@ -1,2 +1,3 @@
 export * from "./service-provider";
 export * as Wallets from "./wallets";
+export { DatabaseInteraction } from "./database-interactions";
