+++ packages/core/bin/config/devnet/app.json
@@ -88,6 +88,9 @@
                 "package": "@arkecosystem/core-logger-pino"
             },
             {
+                "package": "@arkecosystem/core-database"
+            },
+            {
                 "package": "@arkecosystem/core-forger"
             }
         ]
