+++ packages/core/bin/config/mainnet/app.json
@@ -82,6 +82,9 @@
                 "package": "@arkecosystem/core-logger-pino"
             },
             {
+                "package": "@arkecosystem/core-database"
+            },
+            {
                 "package": "@arkecosystem/core-forger"
             }
         ]
